Added ability to ask questions about images
To add the ability to ask the AI assistant about images we had to change both the front end and the back end.

For the front end, Assistant had to be changed to add a button to upload images.  Also, the Assistant component calls the doAsk.js utilitiy and that program had to be changed to create a multimodal form to send to the backend.

In the backend, the /ask route in app.js had to be changed to receive a form and determine whether there are images.  If there are, they had to be moved from the /uploads directory to the /tmp directory.  The names of the images, along with the question and other information had to be sent to the ai_assistant program.

The ai_assistant program had to be changed to accept the images and to format them and attach them to the current thread.  The thread was run with the assistant as normal.